Could not delete file on disk when deleting change #5523
Rejecting or accepting a change for deletion mean the card has to be
deleted. The operation failed because the card was deleted in DB and
then the file was attempted to be deleted on disk but failed because the proxy
was not yet loaded. And loading the proxy with a card without ID made
it generate a query that incorrectly return a lot of records (the entire
change table). This caused memory exhaustion and fatal errored.

So we now ensure to load the proxy before flushing in DB, so the card
still has all needed data to delete the file from disk.