Introduce owner in addition of creator #5557
The creator is purely information for end-user, to be able to know
who originally created something, while the owner is the basis for
acccess rights.

A creator can never change. It is set once for the entire life of the
object. An owner might change, albeit rarely, to "donate" an object
to another user.