Setup controllers to use thesaurus without conflicts
Thesaurus cannot be two way binded because the input format is very different form the outputed one.

We give him an object, and receive a string. To prevent any kind of conflicts, controllers
that use thesaurus need a separated variable for thesaurus input. This variable is out of the model
and need to be initialized smartly when component is loaded and after update().

The output updates the model directly.