Support roman number for centuries in dating #5873
This allow the quizz feature to match dating expressed in roman number
like in the old Dilps. And in addition it also allow to search by
their roman dates (which was not possible in old Dilps).

We also added the error margin in quizz dating matching as it existed
in old Dilps.